backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.11

# Pin bcrypt to avoid backend compatibility issues on some platforms
bcrypt==4.1.2
# Pin passlib for consistent bcrypt backend selection
passlib[bcrypt]==1.7.4

pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv==1.0.1
email-validator
google-auth>=2.35.0
requests>=2.31.0
